[rtc] 移除rtc_core.c/.h 将内容转移到rtc.c/.h
